docs(table_common_stories): replace SortingPaginationExample to examples
replace story in Table to examples folder
