--- src/components/Table/_stories/Table.stories.module.scss
@@ -1,5 +1,0 @@
-.tableHeaderSticky {
-  position: sticky;
-  top: 0;
-  z-index: 2;
-}
