+++ src/components/Table/examples/Example.module.scss
@@ -53,3 +53,9 @@
     border-right: none;
   }
 }
+
+.tableHeaderSticky {
+  position: sticky;
+  top: 0;
+  z-index: 10;
+}
