+++ src/components/Table/examples/index.ts
@@ -13,6 +13,7 @@ export { default as SelectableTableWithPaginationExample } from './SelectableTab
 export { default as SelectableTableWithGroupingHeaderExample } from './SelectableTableWithGroupingHeader';
 export { default as SelectableTableWithVirtualizationExample } from './SelectableTableWithVirtualization';
 export { default as SettingsTableExample } from './SettingsTableExample';
+export { default as SortingPaginationExample } from './CommonTableExample';
 export * from './utils';
 export * from './constants';
 export * from './types';
